IP-165 refactored logo, action buttons, and set flag to remove ktas in timeline view
