+++ src/components/TimeLine/time-line-view.js
@@ -215,6 +215,7 @@ export default class TimeLineView extends Component {
   }
 
   render() {
+    const { areKtaRendered } = this.props;
     let array = [].concat.apply([], Object.values(this.state.dataSplitYears));
 
     const selectY = datum => datum.numberOfActiveProjects;
@@ -289,29 +290,37 @@ export default class TimeLineView extends Component {
           data-intro="Durch diese Ansicht auf <b>Wissenstransferaktivitäten</b> und <b>Drittmittelprojekte</b> wird ermöglicht, beide Elemente des Museums für Naturkunde integrativ und längerfristig zu betrachten."
           data-step="4"
         >
-          {this.renderGridline(lines)}
-          {this.highlightGridLine()}
+          {areKtaRendered && (
+            <>
+              {this.renderGridline(lines)}
+              {this.highlightGridLine()}
+            </>
+          )}
           <div
             data-intro="Im oberen Teil dieser Ansicht werden Wissenstransferaktivitäten gruppiert nach <b>Zielgruppen</b> angezeigt. Die Größe der Kreise deutet die Menge an Aktivitäten mit einer bestimmten Zielgruppe in einem Jahr an. Hierdurch werden längerfristige Perspektiven auf Wissenstransfer ermöglicht."
             data-step="2"
           >
-            <span className={styles.plotTitle}>
-              Wissenstransferaktivitäten <br />
-              <br />
-            </span>
-            <TargetgroupBuckets
-              ktasYearBuckets={this.state.ktasYearBuckets}
-              height={this.state.height * 0.13}
-              width={this.state.width}
-              showYearDetails={this.props.showYearDetails}
-              fullHeight={this.state.height}
-              xScale={year =>
-                xScale(new Date(year.toString()).setHours(0, 0, 0, 0)) +
-                this.state.margin
-              }
-              handleCircleMouseEnter={this.handleCircleMouseEnter}
-              handleCircleMouseLeave={this.handleCircleMouseLeave}
-            />
+            {areKtaRendered && (
+              <>
+                <span className={styles.plotTitle}>
+                  Wissenstransferaktivitäten <br />
+                  <br />
+                </span>
+                <TargetgroupBuckets
+                  ktasYearBuckets={this.state.ktasYearBuckets}
+                  height={this.state.height * 0.13}
+                  width={this.state.width}
+                  showYearDetails={this.props.showYearDetails}
+                  fullHeight={this.state.height}
+                  xScale={year =>
+                    xScale(new Date(year.toString()).setHours(0, 0, 0, 0)) +
+                    this.state.margin
+                  }
+                  handleCircleMouseEnter={this.handleCircleMouseEnter}
+                  handleCircleMouseLeave={this.handleCircleMouseLeave}
+                />
+              </>
+            )}
           </div>
           <SVGWithMargin
             data-intro="Im unteren Teil werden die Anzahl und Laufzeiten von <b>Drittmittelprojekten</b> basierend auf aktuellen Informationen aus dem <a style='color: #afca0b;' href='https://via.museumfuernaturkunde.berlin/wiki/' target='_blank' rel='noopener noreferrer'>VIA-Wiki</a> und gruppiert nach <b>Forschungsgebieten</b> angezeigt. Um die Interpretation von Trend-Entwicklungen zu unterstützen, werden außerdem in grauer Schattierung bisher noch nicht integrierte Daten zu Drittmittelprojekten dargestellt."
@@ -325,9 +334,16 @@ export default class TimeLineView extends Component {
             margin={this.state.margin}
             width={this.state.width}
           >
-            <text fill="#717071" x={-this.state.margin} y="10" fontSize="130%">
-              Forschungsprojekte
-            </text>
+            {areKtaRendered && (
+              <text
+                fill="#717071"
+                x={-this.state.margin}
+                y="10"
+                fontSize="130%"
+              >
+                Forschungsprojekte
+              </text>
+            )}
             {/* a transform style prop to our xAxis to translate it to the bottom of the SVG's content. */}
             <g
               className={styles.xAxis}
@@ -381,6 +397,11 @@ export default class TimeLineView extends Component {
             </g>
           </SVGWithMargin>
           {this.renderProjectsHover()}
+          {!areKtaRendered && (
+            <div className={styles.subScriptLargeScreen}>
+              Anzahl Projekte je Forschungsgebiet pro Jahr
+            </div>
+          )}
         </div>
       </div>
     );
