+++ src/components/TimeLine/time-line-view.module.css
@@ -269,3 +269,8 @@
   color: #717071;
   font-size: 130%;
 }
+
+.subScriptLargeScreen{
+  margin-left: 10px;
+  color: #717071
+}
